tests/wpt/web-platform-tests/html/browsers/history/the-location-interface/non-automated/manual_click_location_replace_during_load-manual.html

<!doctype html>
<title>location.replace with click during load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>The popup blocker must be disabled for this test</p>
<div id="log"></div>
<script>
setup({timeout:3600000});
var t = async_test(undefined, {timeout:3600000});
var win = window.open("manual_click_location_replace_during_load-1.html");

var history_length;
do_test = t.step_func(function(new_length) {
  assert_equals(new_length, history_length);
  t.done();
});
</script>
